requirements.txt
# AWS Data Wrangler 
awswrangler==3.6.0
boto3

# Google Sheets
google-auth==2.29.0
google-auth-oauthlib==1.2.0
gspread==6.0.0

# Airflow 3.0
# apache-airflow==3.0.0
# apache-airflow-providers-amazon==9.0.0
# apache-airflow-providers-google==11.0.0
# apache-airflow-providers-postgres==6.0.0

# dbt
# dbt-core==1.8.0
# dbt-snowflake==1.8.0
# OR dbt-bigquery==1.8.0

#aws wrangler depends on these
sqlalchemy
psycopg2-binary
pg8000>=1.30.0 

# Data processing
pandas==2.1.4
numpy==1.26.3
python-dateutil==2.8.2

# Utilities
python-dotenv==1.0.1
pyyaml==6.0.1

# Testing
pytest==8.0.0
pytest-cov==4.1.0

# Linting
flake8==7.0.0
black==24.1.0
isort==5.13.2